`tools/importer-rest-api-specs`: refactoring the cmds and making logging consistent
